+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- `generate_data_from_vecs` function is only useful for creating histograms. To enable subsampling
+the bitstrings should be saved as they are.
 
 ## [0.0.1] - 2021-03-20
 ### Added